Cleanup + Execution Setup + Have plan for tasking.
Using enkiTS most likely for task scheduling for now. Eventually I'll make my own based on it and implement it in the concurrency module of the core.
I'm still pretty new to lock-preventative multi-threading so I want an good pre-done task scheduler to get the job for me to an extent while I figure out the ropes.
Working on Imgui/UI module now.