GPUVK Backend: Comms milestone reached.
Imgui was abled to work with the hardcoded vulkan backend made for the engine, however it reached its limit with trying to render the model scene properly with the docking feature. The viewports feature outright didn't work right, and in order to fix this I need to essentially modularize the backend for proper dynamic control of rendering on the gpu.

This commit contains the backend working with the new GPUVK_Comms implementation.