Got imgui callback in place, messed with viewport size for testing purposes. Getting ready to start slowly fleshing out rendering module.
In order to render within an imgui viewport window, the scene must be renderered into an offscreen render target. This will need the rendercontext to be redesigned into different variants with at least there being two, one for surface rendering, the other for rendering directly to images and having a copy available on the host process.